+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - Refactored the Advanced Issue Template to V2 with stricter prerequisites and a focus on architectural design (#2016).
 - Refactored the Advanced Issue Template to ensure PR-level quality checklists do not block maintainers during issue creation (#2036)
 - Add automated label sync workflow to propagate labels from linked issues to pull requests (#1716)
+- Added `request-triage-review.yml` and `bot-pr-gfi-review-triage.js` put comment on GFI and Beginner PR to request review.(#1721)
 
 ## [0.2.3] - 2026-03-26